Remove CreateDB.ipynb
This name is confusing, since the file is for ETL of bp and other data.
In the next commit, a new Jupyter Notebook will be added for ETL of
these data with a more appropriate name.

The new notebook will also make better choices about the data -
including renewables and adding relationships.